--- src/by/it/kharitonenko/calc/Vars/Varvars.txt
@@ -1,2 +1,0 @@
-A=5.0
-B={{10.0,15.0,5.0},{5.0,20.0,10.0}}
